requirements.txt
flask==3.0.3
flask-restful==0.3.10
python-dotenv==1.0.1
psycopg2-binary==2.9.9
flask-sqlalchemy==3.1.1